refactored event and metric services
