+++ src/AzureOpenAIProxy.Management/Components/Pages/EventReport.cs
@@ -1,14 +1,13 @@
 ﻿// using AzureOpenAIProxy.Management.Database;
 using AzureOpenAIProxy.Management.Services;
 using Microsoft.AspNetCore.Components;
-// using MudBlazor;
 
 namespace AzureOpenAIProxy.Management.Components.Pages;
 
 public partial class EventReport
 {
     [Inject]
-    private IEventService EventService { get; set; } = null!;
+    private IMetricService MetricService { get; set; } = null!;
 
     [Inject]
     public required IConfiguration Configuration { get; set; }
@@ -19,7 +18,7 @@ public partial class EventReport
 
     protected override async Task OnInitializedAsync()
     {
-        AllEvents = await EventService.GetAllEventsAsync();
+        AllEvents = await MetricService.GetAllEventsAsync();
         // calculate total attendees
         TotalRegistations = AllEvents.Sum(e => e.Registered);
     }
